requirements.txt
asgiref==3.5.2
astroid==2.11.7
atomicwrites==1.4.0
attrs==21.4.0
black==22.6.0
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
Django==4.0.6
django-filter==22.1
django-simple-history==3.1.1
djangorestframework==3.13.1
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.7
mccabe==0.7.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pylint==2.14.4
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-stdnum==1.17
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
tzdata==2022.1
wrapt==1.14.1
